+++ src/app/core/settings.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import * as Store from 'electron-store';
 import { Constants } from './constants';
+import * as os from 'os';
 
 @Injectable()
 export class Settings {
@@ -42,6 +43,15 @@ export class Settings {
         this.settings.set('audioBitrate', v);
     }
 
+    // Custom title bar
+    public get useCustomTitleBar(): boolean {
+        return this.settings.get('useCustomTitleBar');
+    }
+
+    public set useCustomTitleBar(v: boolean) {
+        this.settings.set('useCustomTitleBar', v);
+    }
+
     // FontSize
     public get fontSize(): number {
         return this.settings.get('fontSize');
@@ -74,6 +84,14 @@ export class Settings {
 
         if (!this.settings.has('audioBitrate')) {
             this.settings.set('audioBitrate', 320);
+        }
+
+        if (!this.settings.has('useCustomTitleBar')) {
+            if (os.platform() === 'win32') {
+                this.settings.set('useCustomTitleBar', true);
+            } else {
+                this.settings.set('useCustomTitleBar', false);
+            }
         }
 
         if (!this.settings.has('fontSize')) {